requirements.txt
google-cloud-storage==1.10.0
requests==2.19.1
requests-toolbelt==0.8.0
